Add P42LoggerTestProgram
This commit introduces a test program for the P42Logger with various logging level statements. The program also includes a stopwatch to measure elapsed time. A reference to the P42Logger has been added in the P42LoggerTestProgram's project file and the solution file has been updated to include the new project.